+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Favorites:
   - Correct inline editing
   - Right click on "Favorites" button to edit them
+  - Subfolders in aliases (use "/" or "\" to split as a path)
 
 ## v0.2.8
 - General: